Added Horizontal Scrolling
The layout of the game was turned, so the player now moves up and down as the screen scrolls right to left. Scrolling was added, and we are currently only using two nametables (we really don't need more, they're going to just be black). Added in very rudimentary background graphics for the scrolling.