Began PRNG, cleaned up code
This code introduced a basic algorithm to generate pseudo-random numbers to be used to place asteroids and fuel, as well as to determine their y coordinate. I also added constants for all buttons so you can use BUTTON_START instead of %00010000, for example. Additionally, I changed a lot of the labels in the code to use .done instead of SomeLabelDone, things like that.