Added Collision Handler
Along with some minor code edits and addition of some comments, this commit includes a collision handler. The comments should document it fairly well. Currently, it is technically possible for the algorithm to miss a collision if the player hits the object in the right place, but this will work for now.