Fix Background Collision Code
The code for background collision now actually registers a collision when the player touches an asteroid--before, it was reading the value in a memory address that was not related to the player position. This issue has been fixed, but it must be expanded to check all player tiles.
Additionally, a start pointer was introduced--this could be removed as an optimization, and using a function to add 32 every time to buff_ptr when we want to use it, but for now it stays.