BKG-Player Collisions
The code now checks the whole player for collisions by checking the four tiles of the player using the Y index. Note that the code sometimes misses collisions with the bottom of the player. I have not resolved this yet, but the algorithm seems to work fairly well for now.