Preliminary Game Over Code
Added preliminary code to establish a "Game Over" screen upon colliding with a background object. It isn't perfect and usually also draws asteroids alongside the "GAME OVER" text, though this is not specified in the binary file.
Additionally, the game sometimes "jumps" and causes a game over where there shouldn't be one.
Further, the code for collisions still isn't 100% and usually misses what should be collisions with the bottom half of the player sprite. This is probably due to which pixels/tiles it is checking. Further investigation on all these issues required.