Fixed Game Over Screen
Prior to this commit, the "Game Over" screen would sometimes show a line of asteroids across the screen. Because DrawNewColumn is called in the game over screen draw code, it still runs the loop to draw asteroids on screen, using num_objects. This variable is now set to 0 prior to drawing so that the asteroid draw routine is skipped in DrawNewColumn.